LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointermove_on_chorded_mouse_button-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
PointerMove

Test Description: This test checks if pointermove event are triggered by button state changes
Put your mouse over the black rectangle
Press a button and hold it
Press a second button
Release the second button
Release the first button to complete the test

Pointer Events pointermove on button state changes Tests

The following pointer types were detected: (none).

Refresh the page to run the tests again.


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN pointermove events received for button state changes
